Move baseparser into pip._internal.cli
