+++ src/pip/_internal/basecommand.py
@@ -8,7 +8,7 @@ import os
 import sys
 
 from pip._internal import cmdoptions
-from pip._internal.baseparser import (
+from pip._internal.cli.baseparser import (
     ConfigOptionParser, UpdatingDefaultsHelpFormatter,
 )
 from pip._internal.download import PipSession
