+++ src/pip/_internal/cli/__init__.py
@@ -1,0 +1,4 @@
+"""Subpackage containing all of pip's command line interface related code
+"""
+
+# This file intentionally does not import submodules
